Refactor FRC/FSC module structure and add 3D/accumulation features
- Extract shared code into miplib/analysis/resolution/common.py:
  FRCOptions, _cutoff_correction, namespace_to_frc_options,
  create_fourier_iterator, accumulate_fourier_correlation,
  accumulate_sectioned_fourier_correlation, build_correlation_curve
- Add 3D support to create_fourier_iterator (returns FourierShellIterator)
- Add accumulate_sectioned_fourier_correlation for sectioned shell iterators
- Refactor DirectionalFSC.execute() to use the shared accumulation function
- Fix use_hamming check in calculate_one_image_sectioned_fsc and
  calculate_two_image_sectioned_fsc (was ignoring the option flag)
- Remove empty inline test/ directory in resolution/
- Add tests: 3D iterator, 3D accumulation, sectioned accumulation
- Update all import paths across the codebase
