+++ miplib/analysis/image_quality/image_quality_ranking.py
@@ -1,8 +1,8 @@
 from dataclasses import dataclass
 from pathlib import Path
 
+from miplib.analysis.resolution.common import FRCOptions
 from miplib.analysis.resolution.fourier_ring_correlation import (
-    FRCOptions,
     calculate_single_image_frc,
 )
 from miplib.data.containers.image import Image
