+++ miplib/processing/deconvolution/tracker.py
@@ -4,8 +4,8 @@ import logging
 
 import pandas as pd
 
+from miplib.analysis.resolution.common import FRCOptions
 from miplib.analysis.resolution.fourier_ring_correlation import (
-    FRCOptions,
     calculate_single_image_frc,
 )
 from miplib.data.containers.image import Image
